feat: major update (refactors, design & style, confirm modals, 404 not found page)
