+++ src/components/rooms/SwapDiagram.jsx
