+++ src/main.jsx
@@ -4,9 +4,9 @@ import App from "./App.jsx";
 import AppProvider from "./context/AppContext.jsx";
 
 createRoot(document.getElementById("root")).render(
-  <StrictMode>
+  // <StrictMode>
     <AppProvider>
       <App />
     </AppProvider>
-  </StrictMode>
+  // </StrictMode>
 );
